do some updates
