+++ src/shared/auth/auth.controller.ts
@@ -12,6 +12,8 @@ import { Public, Roles } from '../decorators/roles.decorator';
 import { CurrentUser } from '../decorators/user.decorator';
 import { Role } from '../interface/roles';
 import { OTPService } from './shared/otp.service';
+import { VerifyOtpDto } from './dto/verify-otp.dto';
+import { RegisterDto } from './dto/register.dto';
 
 @Controller('auth')
 export class AuthController {
@@ -21,11 +23,11 @@ export class AuthController {
   @Post('admin/login')
   async superUserLogin(@Body() loginDto: { email: string; password: string }) {
     try {
-      const superUser = await this.authService.validateSuperUser(
+      const { access_token, refresh_token } = await this.authService.validateSuperUser(
         loginDto.email,
         loginDto.password,
       );
-      return new SuccessResponse('Token', { access_token: superUser });
+      return new SuccessResponse('Token', { access_token,refresh_token });
     } catch (err) {
       throw new UnauthorizedException(err?.message, {
         cause: err,
@@ -38,11 +40,11 @@ export class AuthController {
   @Post('user/login')
   async userLogin(@Body() loginDto: { email: string; password: string }) {
     try {
-      const user = await this.authService.validateUserByEmail(
+      const { access_token, refresh_token } = await this.authService.validateUserByEmail(
         loginDto.email,
         loginDto.password,
       );
-      return new SuccessResponse('Token', { access_token: user });
+      return new SuccessResponse('Token', { access_token, refresh_token });
     } catch (err) {
       throw new UnauthorizedException(err?.message, {
         cause: err,
@@ -73,16 +75,39 @@ export class AuthController {
   @Post('user/verify-otp')
   async verifyOtp(@Body() { number, otp }: { number: string; otp: string }) {
     try {
-      const user = await this.authService.validateUserByNumber(
+      const { access_token, refresh_token } = await this.authService.validateUserByNumber(
         number,
         otp,
       );
-      return new SuccessResponse('Token', { access_token: user });
+      return new SuccessResponse('Token', { access_token, refresh_token });
     } catch (err) {
       throw new UnauthorizedException(err?.message, {
         cause: err,
         description: err,
       });
+    }
+  }
+  @Public()
+  @Post('register')
+  async registerUser(@Body() registerDto: RegisterDto) {
+    try {
+      await this.authService.register(registerDto);
+      return new SuccessResponse('OTP sent successfully');
+    } catch (err) {
+      console.log(err);
+
+      return new ErrorResponse();
+    }
+  }
+
+  @Public()
+  @Post('verify')
+  async verifyRegistration(@Body() verifyDto: VerifyOtpDto) {
+    try {
+      const { access_token, refresh_token } = await this.authService.verify(verifyDto.number, verifyDto.otp);
+      return new SuccessResponse('Token', { access_token, refresh_token });
+    } catch (err) {
+      throw new UnauthorizedException(err?.message || 'Verification failed');
     }
   }
 }
