+++ src/shared/auth/auth.service.ts
@@ -1,8 +1,11 @@
-import { Injectable } from '@nestjs/common';
+import { Inject, Injectable, UnauthorizedException } from '@nestjs/common';
 import { PrismaService } from 'src/prisma.service';
 import { OTPService } from './shared/otp.service';
 import { JWTService } from './shared/jwt.service';
 import { BcryptService } from './shared/bcrypt.service';
+import Redis from 'ioredis';
+import { RegisterDto } from './dto/register.dto';
+import { customUUID } from 'src/common/uniqueId.utils';
 
 
 
@@ -12,17 +15,18 @@ export class AuthService {
     private readonly jwtService: JWTService,
     private readonly prisma: PrismaService,
     private readonly otpService: OTPService,
-    private readonly bycrptService: BcryptService
+    private readonly bycrptService: BcryptService,
+    @Inject('REDIS_CLIENT') private readonly redisClient: Redis,
   ) { }
 
 
   async validateSuperUser(email: string, pass: string) {
     const user = await this.prisma.superUser.findFirst({ where: { email } });
     if (user && (await this.bycrptService.comparePassword(pass, user.password))) {
-      return await this.jwtService.createAccessToken({
+      return await this.jwtService.createTokens({
         email: user.email,
-        userId: user.id,
-        role: 'admin',
+        id: user.id,
+        role: 'user',
       });
     }
     throw "Error, couldn't find the user";
@@ -34,9 +38,9 @@ export class AuthService {
     });
     const validPassword = await this.bycrptService.comparePassword(pass, user.password);
     if (user && validPassword) {
-      return await this.jwtService.createAccessToken({
+      return await this.jwtService.createTokens({
         email: user.email,
-        userId: user.id,
+        id: user.id,
         role: 'user',
       });
     }
@@ -51,13 +55,66 @@ export class AuthService {
     });
 
     if (user && await this.otpService.verifyOtp(number, otp)) {
-      return await this.jwtService.createAccessToken({
+      return await this.jwtService.createTokens({
         email: user.email,
-        userId: user.id,
+        id: user.id,
         role: 'user',
       });
     }
     throw new Error("Error, couldn't verify the user with OTP");
+  }
+  async register(registerDto: RegisterDto) {
+    const existingUser = await this.prisma.user.findFirst({
+      where: { OR: [{ email: registerDto.email }, { number: registerDto.number }] },
+    });
+
+    if (existingUser) {
+      throw new UnauthorizedException('Email or Number already in use');
+    }
+
+    const hashedPassword = await this.bycrptService.hashPassword(registerDto.password);
+    const cachedData = { ...registerDto, password: hashedPassword };
+    await this.redisClient.set(`${registerDto.number}`, JSON.stringify(cachedData), 'EX', 300);
+
+    const otp = await this.otpService.generateOtp(registerDto.number);
+    await this.otpService.sendOtpToUser(registerDto.number, otp);
+  }
+
+  async verify(number: string, otp: string) {
+    const isValidOtp = await this.otpService.verifyOtp(number, otp);
+    if (!isValidOtp) {
+      throw new UnauthorizedException('Invalid OTP');
+    }
+
+    const cachedDataString = await this.redisClient.get(`${number}`);
+    if (!cachedDataString) {
+      throw new UnauthorizedException('No registration data found');
+    }
+
+    await this.redisClient.del(`${number}`);
+    
+    const cachedData = JSON.parse(cachedDataString);
+
+    const id = customUUID(20);
+    const user = await this.prisma.user.create({
+      data: {
+        id: id,
+        email: cachedData.email,
+        number: cachedData.number,
+        password: cachedData.password,
+        facebook: cachedData.facebook,
+        instagram: cachedData.instagram,
+        gender: cachedData.gender,
+        type: 'USER',
+        status: 'ACTIVE',
+      },
+    });
+
+    return await this.jwtService.createTokens({
+      email: user.email,
+      id: user.id,
+      role: 'user',
+    });
   }
